+++ libs/common-bls-signing/src/bls_signing.rs
@@ -1,7 +1,7 @@
 use common_bls::{affine::Signature, bls::hash_to_curve};
 use pairing::group::Curve;
 
-use crate::affine::SigningKey;
+use crate::signing_key::SigningKey;
 
 pub fn sign_message(message: &[u8], signing_key: SigningKey) -> Signature {
     let mut proj = hash_to_curve(message);
