+++ libs/common-bls-signing/src/lib.rs
@@ -1,3 +1,3 @@
-pub mod affine;
-pub mod bls;
+pub mod bls_signing;
+pub mod signing_key;
 pub mod threshold;
